requirements.txt
flask>=2.0
rasterio>=1.3
numpy>=1.20
pillow>=9.0
torchgeo>=0.5
